+++ src/datarobot_genai/core/chat/responses.py
@@ -14,13 +14,18 @@
 
 """OpenAI-compatible response helpers for chat interactions."""
 
+import asyncio
+import queue
 import time
 import traceback as tb
 import uuid
 from asyncio import AbstractEventLoop
 from collections.abc import AsyncGenerator
+from collections.abc import AsyncIterator
 from collections.abc import Iterator
 from concurrent.futures import ThreadPoolExecutor
+from typing import Any
+from typing import TypeVar
 
 from openai.types import CompletionUsage
 from openai.types.chat import ChatCompletion
@@ -224,3 +229,41 @@ def to_custom_model_streaming_response_old(
         if last_pipeline_interactions
         else None,
     )
+
+
+T = TypeVar("T")
+
+
+def async_gen_to_sync_thread(
+    async_iterator: AsyncIterator[T],
+    thread_pool_executor: ThreadPoolExecutor,
+    event_loop: asyncio.AbstractEventLoop,
+) -> Iterator[T]:
+    """Run an async iterator in a separate thread and provide a sync iterator."""
+    # A thread-safe queue for communication
+    sync_queue: queue.Queue[Any] = queue.Queue()
+    # A sentinel object to signal the end of the async generator
+    SENTINEL = object()  # noqa: N806
+
+    async def run_async_to_queue() -> None:
+        """Run in the separate thread's event loop."""
+        try:
+            async for item in async_iterator:
+                sync_queue.put(item)
+        except Exception as e:
+            # Put the exception on the queue to be re-raised in the main thread
+            sync_queue.put(e)
+        finally:
+            # Signal the end of iteration
+            sync_queue.put(SENTINEL)
+
+    thread_pool_executor.submit(event_loop.run_until_complete, run_async_to_queue()).result()
+
+    # The main thread consumes items synchronously
+    while True:
+        item = sync_queue.get()
+        if item is SENTINEL:
+            break
+        if isinstance(item, Exception):
+            raise item
+        yield item
